refector!: move all the enum entities to enum directory + separate cases in updateStatus (except reject case)
