+++ src/main/java/com/team2/a2/Facade/ClaimFacade.java
@@ -1,7 +1,7 @@
 package com.team2.a2.Facade;
 
 import com.team2.a2.Model.InsuranceObject.Claim;
-import com.team2.a2.Model.InsuranceObject.ClaimStatus;
+import com.team2.a2.Model.Enum.ClaimStatus;
 import com.team2.a2.Request.InsertClaimRequest;
 
 import java.util.List;
