+++ src/main/java/com/team2/a2/FacadeImpl/ClaimFacadeImpl.java
@@ -2,7 +2,7 @@ package com.team2.a2.FacadeImpl;
 
 import com.team2.a2.Facade.ClaimFacade;
 import com.team2.a2.Model.InsuranceObject.Claim;
-import com.team2.a2.Model.InsuranceObject.ClaimStatus;
+import com.team2.a2.Model.Enum.ClaimStatus;
 import com.team2.a2.Repository.ClaimRepository;
 import com.team2.a2.Request.InsertClaimRequest;
 
@@ -31,6 +31,21 @@ public class ClaimFacadeImpl implements ClaimFacade {
 
     @Override
     public void updateClaimStatus(int id, ClaimStatus status) {
+        Claim claim = claimRepository.getClaimById(id);
+        if (claim == null) return;
+
+        switch (status) {
+            case ACCEPTED:
+                if (claim.getStatus() == ClaimStatus.REJECTED) return;
+
+            case REJECTED:
+                // TODO: Check if we need to check this.
+//                if (claim.getStatus() == ClaimStatus.ACCEPTED) return;
+
+            case PROCESSING:
+                if (claim.getStatus() != ClaimStatus.NEW) return;
+        }
+
         claimRepository.updateClaimStatus(id, status);
     }
