+++ src/main/java/com/team2/a2/InsuranceManagerAllCustomerView.java
@@ -3,9 +3,8 @@ package com.team2.a2;
 import com.team2.a2.Controller.AccountController;
 import com.team2.a2.Model.User.Account;
 import com.team2.a2.Model.User.Customer.Customer;
-import com.team2.a2.Model.User.Customer.CustomerType;
+import com.team2.a2.Model.Enum.CustomerType;
 import com.team2.a2.Model.User.Provider.InsuranceManager;
-import com.team2.a2.Model.User.Provider.InsuranceSurveyor;
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.beans.property.SimpleStringProperty;
