+++ src/main/java/com/team2/a2/InsuranceSurveyorAllCustomerView.java
@@ -1,12 +1,10 @@
 package com.team2.a2;
 
 import com.team2.a2.Controller.AccountController;
-import com.team2.a2.Model.InsuranceObject.Claim;
 import com.team2.a2.Model.User.Account;
 import com.team2.a2.Model.User.Customer.Customer;
-import com.team2.a2.Model.User.Customer.CustomerType;
+import com.team2.a2.Model.Enum.CustomerType;
 import com.team2.a2.Model.User.Provider.InsuranceSurveyor;
-import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.beans.property.SimpleStringProperty;
@@ -20,15 +18,10 @@ import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
-import javafx.scene.control.TextField;
 import javafx.stage.Stage;
 
 import java.io.IOException;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.net.URL;
-import java.util.Date;
-import java.util.Objects;
 import java.util.ResourceBundle;
 
 public class InsuranceSurveyorAllCustomerView implements Initializable {
