+++ src/main/java/com/team2/a2/Model/Enum/ClaimStatus.java
@@ -1,4 +1,4 @@
-package com.team2.a2.Model.InsuranceObject;
+package com.team2.a2.Model.Enum;
 
 public enum ClaimStatus {
     NEW,
