+++ src/main/java/com/team2/a2/Model/Enum/CustomerType.java
@@ -1,4 +1,4 @@
-package com.team2.a2.Model.User.Customer;
+package com.team2.a2.Model.Enum;
 
 public enum CustomerType {
     POLICY_HOLDER,
