+++ src/main/java/com/team2/a2/Repository/AccountRepository.java
@@ -3,7 +3,6 @@ package com.team2.a2.Repository;
 import com.team2.a2.ConnectionManager;
 import com.team2.a2.Model.Enum.AccountType;
 import com.team2.a2.Model.User.Account;
-import com.team2.a2.Model.User.Customer.CustomerType;
 import com.team2.a2.Request.UpdateAccountRequest;
 
 import java.sql.*;
