+++ src/main/java/com/team2/a2/Repository/ClaimRepository.java
@@ -2,7 +2,7 @@ package com.team2.a2.Repository;
 
 import com.team2.a2.ConnectionManager;
 import com.team2.a2.Model.InsuranceObject.Claim;
-import com.team2.a2.Model.InsuranceObject.ClaimStatus;
+import com.team2.a2.Model.Enum.ClaimStatus;
 
 import java.sql.*;
 import java.util.ArrayList;
