+++ src/main/java/com/team2/a2/Repository/InsuranceCardRepository.java
@@ -2,10 +2,6 @@ package com.team2.a2.Repository;
 
 import com.team2.a2.ConnectionManager;
 import com.team2.a2.Model.InsuranceObject.InsuranceCard;
-import com.team2.a2.Model.User.Admin;
-import com.team2.a2.Model.User.Customer.Customer;
-import com.team2.a2.Model.User.Customer.CustomerType;
-import com.team2.a2.Request.InsertCustomerRequest;
 import com.team2.a2.Request.InsertInsuranceCardRequest;
 
 import java.sql.*;
