+++ src/main/java/com/team2/a2/Request/InsertClaimRequest.java
@@ -1,6 +1,6 @@
 package com.team2.a2.Request;
 
-import com.team2.a2.Model.InsuranceObject.ClaimStatus;
+import com.team2.a2.Model.Enum.ClaimStatus;
 
 import java.sql.Date;
